requirements.txt
<<<<<<< HEAD
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
=======
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
celery>=5.0,<6.0
django-environ>=0.9,<1.0
drf-yasg>=1.21,<2.0
mysqlclient>=2.1,<3.0
rabbitmq-server  # Note: This is usually installed via system package manager
>>>>>>> 797fbbb7 (Complete project setup with MySQL, Swagger, and REST API configuration)
